requirements.txt
requests
python-dotenv
flask-security-too
flask-migrate>=3.0.0
flask-debugtoolbar
flask-jwt-extended>=4.0.0
gevent-websocket
PyMySQL>=0.9.3
SQLAlchemy
Flask-SQLAlchemy>=2.3.2
gevent>=1.4.0
gunicorn>=19.9.0
eventlet
flower<=0.9.7
flask>=2.0.0
Werkzeug>=2.0.0
itsdangerous>=2.0.0
mysqlclient
concurrent-log-handler
flask-sugar
flask-admin
pydantic
shortuuid
aliyun-python-sdk-core
aliyun-python-sdk-core-v3
kubernetes==19.15.0